project layer restructure
