+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/controller/ScheduleController.java
@@ -1,5 +1,6 @@
-package com.udacity.jdnd.course3.critter.schedule;
+package com.udacity.jdnd.course3.critter.controller;
 
+import com.udacity.jdnd.course3.critter.dto.ScheduleDTO;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
