+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/EmployeeRequestDTO.java
@@ -1,4 +1,6 @@
-package com.udacity.jdnd.course3.critter.user;
+package com.udacity.jdnd.course3.critter.dto;
+
+import com.udacity.jdnd.course3.critter.model.EmployeeSkill;
 
 import java.time.LocalDate;
 import java.util.Set;
