+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/PetDTO.java
@@ -1,4 +1,6 @@
-package com.udacity.jdnd.course3.critter.pet;
+package com.udacity.jdnd.course3.critter.dto;
+
+import com.udacity.jdnd.course3.critter.model.PetType;
 
 import java.time.LocalDate;
