+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/model/Customer.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.entity;
+package com.udacity.jdnd.course3.critter.model;
 
 import org.hibernate.annotations.Nationalized;
