+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/model/EmployeeSkill.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.user;
+package com.udacity.jdnd.course3.critter.model;
 
 /**
  * A example list of employee skills that could be included on an employee or a schedule request.
