+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/model/Pet.java
@@ -1,6 +1,5 @@
-package com.udacity.jdnd.course3.critter.entity;
+package com.udacity.jdnd.course3.critter.model;
 
-import com.udacity.jdnd.course3.critter.pet.PetType;
 import org.hibernate.annotations.Nationalized;
 
 import javax.persistence.*;
