+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/model/PetType.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.pet;
+package com.udacity.jdnd.course3.critter.model;
 
 /**
  * A example list of pet type metadata that could be included on a request to create a pet.
